# Optimize AES-CTR encryption/decryption