--- a/internal/aesctr/export_test.go
+++ b/internal/aesctr/export_test.go
@@ -1,0 +1,10 @@
+// Copyright 2018 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package aesctr
+
+// Export internal functions for testing.
+var XorBytes = xorBytes
+var AnyOverlap = anyOverlap
+var InexactOverlap = inexactOverlap