--- a/srtcp.go
+++ b/srtcp.go
@@ -1,7 +1,6 @@
 package srtp
 
 import (
-	"crypto/cipher"
 	"encoding/binary"
 
 	"github.com/pions/rtcp"
@@ -23,8 +22,9 @@ func (c *Context) decryptRTCP(dst, encrypted []byte) ([]byte, error) {
 	index := binary.BigEndian.Uint32(srtcpIndexBuffer) &^ (1 << 31)
 	ssrc := binary.BigEndian.Uint32(encrypted[4:])
 
-	stream := cipher.NewCTR(c.srtcpBlock, c.generateCounter(uint16(index&0xffff), index>>16, ssrc, c.srtcpSessionSalt))
-	stream.XORKeyStream(out[8:], out[8:])
+	counter := c.generateCounter(uint16(index&0xffff), index>>16, ssrc, c.srtcpSessionSalt)
+	c.srtcpStream.Reset(counter)
+	c.srtcpStream.XORKeyStream(out[8:], out[8:])
 
 	return out, nil
 }
@@ -53,8 +53,9 @@ func (c *Context) encryptRTCP(dst, decrypted []byte) ([]byte, error) {
 	}
 
 	// Encrypt everything after header
-	stream := cipher.NewCTR(c.srtcpBlock, c.generateCounter(uint16(c.srtcpIndex&0xffff), c.srtcpIndex>>16, ssrc, c.srtcpSessionSalt))
-	stream.XORKeyStream(out[8:], out[8:])
+	counter := c.generateCounter(uint16(c.srtcpIndex&0xffff), c.srtcpIndex>>16, ssrc, c.srtcpSessionSalt)
+	c.srtcpStream.Reset(counter)
+	c.srtcpStream.XORKeyStream(out[8:], out[8:])
 
 	// Add SRTCP Index and set Encryption bit
 	out = append(out, make([]byte, 4)...)