--- a/srtp_test.go
+++ b/srtp_test.go
@@ -292,12 +292,13 @@ func BenchmarkEncryptRTP(b *testing.B) {
 		b.Fatal(err)
 	}
 
-	pkt := &rtp.Packet{Payload: make([]byte, 100)}
+	pkt := &rtp.Packet{Payload: make([]byte, 1000)}
 	pktRaw, err := pkt.Marshal()
 	if err != nil {
 		b.Fatal(err)
 	}
 
+	b.SetBytes(int64(len(pktRaw)))
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
@@ -314,7 +315,7 @@ func BenchmarkEncryptRTPInPlace(b *testing.B) {
 		b.Fatal(err)
 	}
 
-	pkt := &rtp.Packet{Payload: make([]byte, 100)}
+	pkt := &rtp.Packet{Payload: make([]byte, 1000)}
 	pktRaw, err := pkt.Marshal()
 	if err != nil {
 		b.Fatal(err)
@@ -322,6 +323,7 @@ func BenchmarkEncryptRTPInPlace(b *testing.B) {
 
 	buf := make([]byte, 0, len(pktRaw)+10)
 
+	b.SetBytes(int64(len(pktRaw)))
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
@@ -333,30 +335,27 @@ func BenchmarkEncryptRTPInPlace(b *testing.B) {
 }
 
 func BenchmarkDecryptRTP(b *testing.B) {
-	sequenceNumber := uint16(5000)
-	encrypted := []byte{0x6d, 0xd3, 0x7e, 0xd5, 0x99, 0xb7, 0x2d, 0x28, 0xb1, 0xf3, 0xa1, 0xf0, 0xc, 0xfb, 0xfd, 0x8}
-
-	encryptedPkt := &rtp.Packet{
-		Payload: encrypted,
-		Header: rtp.Header{
-			SequenceNumber: sequenceNumber,
-		},
-	}
-
-	encryptedRaw, err := encryptedPkt.Marshal()
-	if err != nil {
-		b.Fatal(err)
-	}
-
 	context, err := buildTestContext()
 	if err != nil {
 		b.Fatal(err)
 	}
 
+	pkt := &rtp.Packet{Payload: make([]byte, 1000)}
+	pktRaw, err := pkt.Marshal()
+	if err != nil {
+		b.Fatal(err)
+	}
+
+	buf, err := context.EncryptRTP(nil, pktRaw, nil)
+	if err != nil {
+		b.Fatal(err)
+	}
+
+	b.SetBytes(int64(len(buf)))
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
-		_, err := context.DecryptRTP(nil, encryptedRaw, nil)
+		_, err := context.DecryptRTP(nil, buf, nil)
 		if err != nil {
 			b.Fatal(err)
 		}