@kixelated 2019-03-23
Let me know if you would prefer that I just fork `pions/srtp` instead this approach.